requirements.txt
numpy==1.20.1
cython
scikit-surprise==1.1.1
scikit-learn==0.24.1
jupyter==1.0.0
pandas==1.2.3
scikit-learn==0.24.1
matplotlib==3.3.4